+++ nflow-explorer/src/app/config/index.js
@@ -4,6 +4,7 @@
   angular.module('nflowExplorer.config', [
     'nflowExplorer.config.console',
     'nflowExplorer.config.routes',
+    'nflowExplorer.config.notificationFilter',
   ])
   .constant('config', new Config());